# stat/distuv: Improve Gamma.Rand performance in alpha >1 case