Please take a look.

Fixes #1453.

benchstat (from the new GammaRand benchmark):
```
name              old time/op  new time/op  delta
GammaRand/case_0  90.3ns ± 5%  91.1ns ± 3%     ~     (p=0.152 n=10+8)
GammaRand/case_1   215ns ± 4%   213ns ± 3%     ~     (p=0.238 n=10+10)
GammaRand/case_2  25.5ns ± 3%  25.7ns ± 3%     ~     (p=0.534 n=9+10)
GammaRand/case_3  72.4ns ± 4%  48.5ns ± 5%  -32.97%  (p=0.000 n=9+10)
GammaRand/case_4  72.0ns ± 6%  47.1ns ± 6%  -34.50%  (p=0.000 n=10+10)
```

<!--
Checklist:

- API changes have been discussed
- code is goformated correctly (goimports)
- packages with generated code have had code generation run
- tests pass locally
- linked to relevant issues

Please make sure your commit message summary line and pull request
title match the Go convention; a one-line summary of the change,
prefixed by the primary affected package that should complete the
sentence, "This change modifies Gonum to _____."
-->