@romanwerpachowski 2020-09-15
Thanks. Can you also add the benchmark from the issue to the PR?