@sbinet 2020-09-16
thanks as well for the nice improvement.

we are a bit nit-picky (sorry about that), but could you remove the "A+C" authorship modification from this PR and send another one with only the authorship bits in?
this way, we can have a nice commit log, say:
```
stat/distuv: improve Gamma.Rand performances
A+C: add Mark Canning
```

we do like nice commit logs :) (like the ones from the Go project itself. at least that's the goal)

thanks again.