@sbinet 2020-09-16
(btw, I see you have a nice fft package, faster than the one from scientificgo.org. did you benchmark it against gonum's?
well, we could discuss this on gonum-dev instead of hijacking that PR)