@argusdusty 2020-09-16
I've removed the new contributor stuff. I'll send out another pull request containing it after this one. 

Regarding my FFT library, I did benchmark it against gonum/dsp (that would be the "BenchmarkGoDSPFFT" benchmarks), it's significantly faster, but as gonum/dsp is a multithreaded library and mine is single-threaded, it's not a drop-in replacement.