@kortschak 2020-09-16
@argusdusty Our FFT implementation isn't multi-threaded. The current implementation is in an internal package so that we can swap it out when a better one becomes available, keeping the API provided in dsp/fourier the same. The requirements are (and this can be done piecemeal); real sequence FFT, complex sequence FFT, DCT, DST and QW FFT. The current implementation is sequence-length agnostic (no 2^n length requirement) and that behaviour should be kept.