@argusdusty 2020-09-16
Apparently I got go-dsp mixed up with gonum/dsp. I added gonum's fft to my benchmarks and my code still runs a few times faster, but only supports 2^n length inputs, so still not a drop-in replacement.