@kortschak 2020-09-16
Continue off-topic: I just went back and had another look and NumPy uses the same algorithm as we do, but in C, being a translation of FFTPACK's code. Maybe it's just a matter of looking at where we are losing performance.