@argusdusty 2020-09-17
numpy's fft is actually rather slow (though some of that is probably from type conversion/python overhead). That's part of why I made my fft library. fftw is the best, but it's not easy to use from Go/Python without C wrappers, and the code is very complex and hard to translate. Going by http://www.fftw.org/speed/Pentium4-3.06GHz/ fftw runs ~3-4x faster than fftpack, which sounds about right.