@sbinet 2020-09-17
needs https://github.com/gonum/gonum/pull/1460